Added willPaginate setting (will rename it) for Backbone.js friendly pagination based on https://gist.github.com/886863
Actually, that gist is very outdated. But the idea is that, when sending paginated data to a backbone app, you alter your data to:

    { 'models': [] //models here,
      'page': x,
      'per_page': x,
      'total': x
    }

This is inspired by willPaginate (a Ruby gem) and it used actively in my own JS applications. I want to allow people to map their own pagination values and choose the key for 'models' instead of me forcing my own conventions. This is comming in the next few commits.

When willPaginate is non-false, any model arrays will be shifted out of their tableized name key, if it is available.

eg. if you extracted customers.{n}.Customer to customers, it will shift it all out of customers. This makes it more backbone friendly. I will move this out to a more friendly variable soon too.