Large refactor of plugin to extend RequestHandlerComponent
Removed _modelizePost from initialize and remove method entirely.

I've used my own method from another project as it suited me better. It is also only called on JSON POST/PUT data as the parent RequestHandlerComponent class handles XML data. The unrootModelData method moves lowercase keys to a model key but leaves uppercase (likely to be nested models) in place.

Removed headers method

I can't see the need once I've refactored it to properly take advantage of the parent RequestHandlerComponent

Renamed ::$isActive to ::$isRestful because isActive seems a little vague whereas isRestful implies that WE THE COMPONENT are taking charge because it's restful. Makes more sense

Removing beforeRedirect completely in favour of the parent RequestHandlerComponent::beforeRedirect as it resets any POST data and then calls requestAction allowing you to, for example, redirect an AJAX POST/PUT to a GET and reduce any complex logic needed to return an updated model back to the client.

References to request->ext are replaced with $this->ext. The parent RequestHandlerComponent already imports the detected extension in RequestHandlerComponent::initialize() and I have added my own override to force a restful response for an extensionless ajax request.

::credentials() is now a fully fledged Class of its own. RestCredentials. The way it was managed (with multiple function arguments triggering different functions) looked like a Javascript developer making an object. So I made it in to an object. This way, you initialize it during component construction and then, after that, you call get(). Makes it better for writing tests.

Next, I'm going to sort out the logging stuff. There seems to be a lot of funky log activity which could be refactored by putting it in to the RestLog model constructor instead of, again, a strange pseudo-override log method