Added tiny hack to RestComponent::inject to use {n} on the RHS
I really wanted to be able to achieve extract rules like this:

    'customers.{n}.Customer' => 'customers',
    'customers.{n}.Foo' => 'customers.{n}.Foo'

This would allow me to put the customer data at the root with nested model data. It's a niche case but it allows me to reduce AJAX requests in specific instances for trivial pieces of hasOne/belongsTo data.